requirements.txt
Django==5.2.8
djangorestframework
channels
mysql-connector-python
daphne
django-cors-headers
anthropic
python-pptx
reportlab
Pillow
